requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
cffi==1.12.3
Click==7.0
cryptography==2.7
decorator==4.4.0
Django==2.2.3
django-cors-headers==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.10.2
entrypoints==0.3
flake8==3.7.8
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.0.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
zipp==0.5.2
